EDIT: Renamed LoginPage to LoginRegPage so that it renders both LoginPage and Registration page depending on the prop displayLoginPage. Also changed some css so that registration page is consistent with login page
