+++ frontend/src/CSS/registrationForm.css
@@ -1,20 +1,10 @@
 .registration-form-container {
-    display: flex;
-    flex-direction: column;
-    width: 60%;
-    align-items: center;
-    margin-left: 20%;
-}
-
-.registration-form-input-container {
-    width: 60%;
-    
+    padding: 30px 80px;
+    text-align: center;
+    width: 100%;
+    border: black solid 2px;
 }
 
 p {
     margin-top: 10px;
-}
-
-.registration-form-button {
-    float: right;
 }
